[fix] 미션 글쓰기 그룹 선택 터치 비활성화

+++ app/src/main/java/com/dnd_8th_4_android/wery/presentation/ui/mission/mymission/adapter/MissionCardAdapter.kt
@@ -42,7 +42,7 @@ class MissionCardAdapter() :
             binding.data = data
             binding.root.setOnClickListener {
                 val intent = Intent(it.context, MissionDetailActivity::class.java)
-                intent.putExtra("missionId",data.missionId)
+                intent.putExtra("missionId",data.missionId.toInt())
                 (it.context).startActivity(intent)
             }
         }
